fix: card move dialog auto close
