+++ src/App.vue
@@ -41,6 +41,7 @@
 			<router-view name="sidebar" :visible="!cardDetailsInModal || !$route.params.cardId" />
 		</div>
 		<KeyboardShortcuts />
+		<CardMoveDialog />
 	</NcContent>
 </template>
 
@@ -52,12 +53,14 @@ import { NcModal, NcContent, NcAppContent } from '@nextcloud/vue'
 import { BoardApi } from './services/BoardApi.js'
 import { emit, subscribe } from '@nextcloud/event-bus'
 import { loadState } from '@nextcloud/initial-state'
+import CardMoveDialog from './CardMoveDialog.vue'
 
 const boardApi = new BoardApi()
 
 export default {
 	name: 'App',
 	components: {
+		CardMoveDialog,
 		AppNavigation,
 		NcModal,
 		NcContent,
